server/requirements.txt
asgiref==3.7.2
awsebcli==3.20.10
blessed==1.20.0
cement==2.8.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.3
Django==5.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
Markdown==3.5.1
packaging==23.2
pathspec==0.10.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.8.0
urllib3==1.26.18
wcwidth==0.1.9
